+++ src/Lemmon/StringValidator.php
@@ -4,6 +4,7 @@ namespace Lemmon;
 
 class StringValidator extends FieldValidator
 {
+    use OneOfTrait;
     /**
      * @inheritDoc
      */
@@ -33,7 +34,7 @@ class StringValidator extends FieldValidator
 
     public function email(?string $message = null): static
     {
-        return $this->addValidation(
+        return $this->satisfies(
             fn ($value, $key = null, $input = null) => filter_var($value, FILTER_VALIDATE_EMAIL) !== false,
             $message ?? 'Value must be a valid email address'
         );
@@ -41,7 +42,7 @@ class StringValidator extends FieldValidator
 
     public function url(?string $message = null): static
     {
-        return $this->addValidation(
+        return $this->satisfies(
             fn ($value, $key = null, $input = null) => filter_var($value, FILTER_VALIDATE_URL) !== false,
             $message ?? 'Value must be a valid URL'
         );
@@ -49,7 +50,7 @@ class StringValidator extends FieldValidator
 
     public function uuid(?string $message = null): static
     {
-        return $this->addValidation(
+        return $this->satisfies(
             fn ($value, $key = null, $input = null) => preg_match('/^[0-9a-f]{8}-[0-9a-f]{4}-[1-5][0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$/i', $value) === 1,
             $message ?? 'Value must be a valid UUID'
         );
@@ -57,7 +58,7 @@ class StringValidator extends FieldValidator
 
     public function ip(?string $message = null): static
     {
-        return $this->addValidation(
+        return $this->satisfies(
             fn ($value, $key = null, $input = null) => filter_var($value, FILTER_VALIDATE_IP) !== false,
             $message ?? 'Value must be a valid IP address'
         );
@@ -65,7 +66,7 @@ class StringValidator extends FieldValidator
 
     public function minLength(int $min, ?string $message = null): static
     {
-        return $this->addValidation(
+        return $this->satisfies(
             fn ($value, $key = null, $input = null) => mb_strlen($value) >= $min,
             $message ?? "Value must be at least {$min} characters long"
         );
@@ -73,7 +74,7 @@ class StringValidator extends FieldValidator
 
     public function maxLength(int $max, ?string $message = null): static
     {
-        return $this->addValidation(
+        return $this->satisfies(
             fn ($value, $key = null, $input = null) => mb_strlen($value) <= $max,
             $message ?? "Value must be at most {$max} characters long"
         );
@@ -81,7 +82,7 @@ class StringValidator extends FieldValidator
 
     public function length(int $exact, ?string $message = null): static
     {
-        return $this->addValidation(
+        return $this->satisfies(
             fn ($value, $key = null, $input = null) => mb_strlen($value) === $exact,
             $message ?? "Value must be exactly {$exact} characters long"
         );
@@ -89,7 +90,7 @@ class StringValidator extends FieldValidator
 
     public function pattern(string $regex, ?string $message = null): static
     {
-        return $this->addValidation(
+        return $this->satisfies(
             fn ($value, $key = null, $input = null) => preg_match($regex, $value) === 1,
             $message ?? 'Value does not match the required pattern'
         );
@@ -97,7 +98,7 @@ class StringValidator extends FieldValidator
 
     public function datetime(string $format = 'Y-m-d\TH:i:s', ?string $message = null): static
     {
-        return $this->addValidation(
+        return $this->satisfies(
             function ($value, $key = null, $input = null) use ($format) {
                 $date = \DateTime::createFromFormat($format, $value);
                 return $date !== false && $date->format($format) === $value;
@@ -108,7 +109,7 @@ class StringValidator extends FieldValidator
 
     public function date(string $format = 'Y-m-d', ?string $message = null): static
     {
-        return $this->addValidation(
+        return $this->satisfies(
             function ($value, $key = null, $input = null) use ($format) {
                 $date = \DateTime::createFromFormat($format, $value);
                 return $date !== false && $date->format($format) === $value;